Corrigido update cliente

+++ src/main/java/com/vinizorza/seguradoraveiculo/repository/ClienteRepository.java
@@ -3,5 +3,5 @@ package com.vinizorza.seguradoraveiculo.repository;
 import com.vinizorza.seguradoraveiculo.model.Cliente;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ClienteRepository extends MongoRepository<Cliente, Integer> {
+public interface ClienteRepository extends MongoRepository<Cliente, Long> {
 }
